* DataLayer/ISequenceCollection.h (count): Return size_t not int. * DataLayer/SequenceCollectionHash.cpp: Ditto. * DataLayer/SequenceCollectionHash.h: Ditto. * Parallel/NetworkSequenceCollection.cpp: Ditto. * Parallel/NetworkSequenceCollection.h: Ditto.
git-svn-id: https://svn.bcgsc.ca/svn/abyss/trunk@46152 ad02020f-2a26-4133-947e-1829317c1416

+++ DataLayer/ISequenceCollection.h
@@ -59,7 +59,7 @@ class ISequenceCollection
 		virtual int getMultiplicity(const PackedSeq& seq) = 0;
 		
 		// Return the number of sequences in the collection
-		virtual int count() const = 0;
+		virtual size_t count() const = 0;
 		
 		// Clear the specified flag from every sequence in the collection
 		virtual void wipeFlag(SeqFlag flag) = 0;
